Brought back chat component, minor changes.
