+++ interlecture/playground/consumers.py
@@ -2,6 +2,7 @@ from channels import Group
 
 def ws_add(message):
   message.reply_channel.send({"accept": True})
+  message.reply_channel.send({"text": "Welcome"})
   Group("chat").add(message.reply_channel)
 
 def ws_message(message):
